perf: 日志模块缓冲MQ队列机制优化 #3089
