+++ src/backend/ci/core/common/common-stream/src/main/kotlin/com/tencent/devops/common/stream/pulsar/integration/PulsarProducerMessageHandler.kt
@@ -25,25 +25,31 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.stream.pulsar.integration.outbound
+package com.tencent.devops.common.stream.pulsar.integration
 
-import com.tencent.devops.common.stream.pulsar.properties.PulsarBinderConfigurationProperties
+import com.tencent.devops.common.stream.pulsar.constant.Serialization
+import com.tencent.devops.common.stream.pulsar.properties.PulsarProperties
 import com.tencent.devops.common.stream.pulsar.properties.PulsarProducerProperties
 import com.tencent.devops.common.stream.pulsar.support.PulsarMessageConverterSupport
 import com.tencent.devops.common.stream.pulsar.util.PulsarUtils
+import com.tencent.devops.common.stream.pulsar.util.SchemaUtils
+import org.apache.pulsar.client.api.CompressionType
+import org.apache.pulsar.client.api.HashingScheme
+import org.apache.pulsar.client.api.MessageRoutingMode
 import org.apache.pulsar.client.api.Producer
-import org.apache.pulsar.client.api.PulsarClient
+import org.apache.pulsar.client.api.ProducerCryptoFailureAction
 import org.apache.pulsar.client.api.TypedMessageBuilder
+import org.springframework.cloud.stream.binder.ExtendedProducerProperties
 import org.springframework.cloud.stream.provisioning.ProducerDestination
 import org.springframework.context.Lifecycle
 import org.springframework.integration.handler.AbstractMessageHandler
 import org.springframework.messaging.Message
+import java.util.concurrent.TimeUnit
 
 class PulsarProducerMessageHandler(
     private val destination: ProducerDestination,
-    private val producerProperties: PulsarProducerProperties? = null,
-    private val pulsarClient: PulsarClient,
-    private val pulsarProperties: PulsarBinderConfigurationProperties
+    private val producerProperties: ExtendedProducerProperties<PulsarProducerProperties>? = null,
+    private val pulsarProperties: PulsarProperties
 ) : AbstractMessageHandler(), Lifecycle {
     private var producer: Producer<Any>? = null
 
@@ -61,10 +67,10 @@ class PulsarProducerMessageHandler(
             namespace = pulsarProperties.namespace,
             topic = destination.name
         )
-        producer = PulsarProducerFactory.initPulsarProducer(
+        producer = generatePulsarProducer(
             topic = topic,
-            producerProperties = producerProperties,
-            pulsarClient = pulsarClient
+            pulsarProperties = pulsarProperties,
+            producerProperties = producerProperties
         )
     }
 
@@ -103,5 +109,36 @@ class PulsarProducerMessageHandler(
 
     override fun isRunning(): Boolean {
         return running
+    }
+
+    private fun generatePulsarProducer(
+        topic: String,
+        pulsarProperties: PulsarProperties,
+        producerProperties: ExtendedProducerProperties<PulsarProducerProperties>
+    ): Producer<Any> {
+        val pulsarProducerProperties = producerProperties.extension
+        with(pulsarProducerProperties) {
+            val builder = PulsarUtils.getClientBuilder(pulsarProperties)
+
+            // TODO 消息序列化方式需要调整， producer需要缓存
+            val producer = builder.build().newProducer(
+                SchemaUtils.getSchema(Serialization.valueOf(serialType), serialClass)
+            ).topic(topic)
+            if (!producerName.isNullOrBlank()) {
+                producer.producerName(producerName)
+            }
+            producer.sendTimeout(sendTimeoutMs, TimeUnit.MILLISECONDS)
+                .blockIfQueueFull(blockIfQueueFull)
+                .maxPendingMessages(maxPendingMessages)
+                .maxPendingMessagesAcrossPartitions(maxPendingMessagesAcrossPartitions)
+                .messageRoutingMode(MessageRoutingMode.valueOf(messageRoutingMode))
+                .hashingScheme(HashingScheme.valueOf(hashingScheme))
+                .cryptoFailureAction(ProducerCryptoFailureAction.valueOf(cryptoFailureAction))
+                .batchingMaxPublishDelay(batchingMaxPublishDelayMicros, TimeUnit.MILLISECONDS)
+                .batchingMaxMessages(batchingMaxMessages)
+                .enableBatching(batchingEnabled)
+                .compressionType(CompressionType.valueOf(compressionType))
+            return producer.create() as Producer<Any>
+        }
     }
 }
