+++ src/backend/ci/core/common/common-stream/src/main/kotlin/com/tencent/devops/common/stream/pulsar/properties/PulsarConsumerProperties.kt
@@ -47,6 +47,10 @@ data class PulsarConsumerProperties(
      * 	Key_Shared
      */
     var subscriptionType: String = "Shared",
+
+    var numIoThreads: Int? = null,
+    var numListenerThreads: Int? = null,
+    var connectionsPerBroker: Int? = null,
     /**
      * Size of a consumer's receiver queue.
      * For example, the number of messages accumulated by a consumer before an application calls Receive.
